refactor: AddModal 분 및 메모이제이션 적용
