+++ src/client/src/setting/channel/permission/components/FolderMemberRow.jsx
@@ -59,4 +59,4 @@ FolderMemberRow.defaultProps = {
   code: '',
 };
 
-export default FolderMemberRow;
+export default React.memo(FolderMemberRow);
